backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
openai==1.0.0
boto3==1.26.135
gunicorn==20.1.0
pytest==7.3.1
pytest-cov==4.1.0
werkzeug==2.2.3
